Separate scripted simulator data from runner logic
